data-ingestion/requirements.txt
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
apscheduler==3.10.4
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
python-Levenshtein==0.23.0
